backend/requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
pandas
numpy
cloudinary